Upgrade spring test to Spring Boot 3.3.1
This commit also introduces several refinements:
- Upgrade to Java 21 with Liberica JRE
- Make Message a top level class
- Properly exclude autoconfigurations based on profiles
- Use Spring Boot 3.3 executable JAR unpack feature
- Remove java options enabled by default
- Keep debug symbol in JVM bytecode for parameter name resolution
