+++ frameworks/Java/spring/src/main/java/hello/model/Message.java
@@ -1,0 +1,15 @@
+package hello.model;
+
+public class Message {
+
+	private final String message;
+
+	public Message(String message) {
+		this.message = message;
+	}
+
+	public String getMessage() {
+		return message;
+	}
+
+}
