+++ frameworks/Java/spring/src/main/resources/application.yml
@@ -1,10 +1,15 @@
 ---
 spring:
-  threads:
-    virtual:
-      enabled: true
   jpa:
     open-in-view: false
+  config:
+    activate:
+      on-profile: jdbc
+  autoconfigure:
+    exclude: org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration,org.springframework.boot.autoconfigure.data.jpa.JpaRepositoriesAutoConfiguration,org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration,org.springframework.boot.autoconfigure.data.mongo.MongoRepositoriesAutoConfiguration
+
+---
+spring:
   config:
     activate:
       on-profile: jdbc,jpa
@@ -25,6 +30,8 @@ spring:
   config:
     activate:
       on-profile: jpa
+  autoconfigure:
+    exclude: org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration,org.springframework.boot.autoconfigure.data.mongo.MongoRepositoriesAutoConfiguration
   jpa:
     database-platform: org.hibernate.dialect.PostgreSQLDialect
 
@@ -33,6 +40,8 @@ spring:
   config:
     activate:
       on-profile: mongo
+  autoconfigure:
+    exclude: org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration,org.springframework.boot.autoconfigure.data.jpa.JpaRepositoriesAutoConfiguration,org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
 
 spring.data.mongodb:
   host: tfb-database
